Detect Katsuna OS through build.prop (1/2)
KatsunaUpdater is an app which can be uninstalled or frozen in a rooted system.
It can also be removed by us if we ever move to an in-house solution (KatsunaUpdater2).

Detect Katsuna OS by reading for the ro.katsuna.version prop, which will only be set
during build time in a Katsuna OS system (it can't be modified in runtime).

Also check commit in KatsunaLauncher

This was tested on Katsuna OS and on a OEM stock Android image